+++ django_mapengine/static/django_mapengine/js/popup.js
@@ -1,4 +1,7 @@
 
+let popupClicked = false;
+const POPUP_DOUBLE_CLICK_TIMEOUT = 300;
+
 PubSub.subscribe(mapEvent.MAP_SOURCES_LOADED, add_popups);
 
 
@@ -18,8 +21,6 @@ function createListByName(name, series) {
 }
 
 function checkPop(layerID) {
-  console.log(layerID)
-
   if (!(layerID in map_store.cold.popups)) return false;
   if (map_store.cold.currentChoropleth === null) {
     return map_store.cold.popups[layerID].atDefaultLayer;
@@ -38,6 +39,11 @@ function add_popups() {
 
 function add_popup(layerID) {
   map.on("click", layerID, function (event) {
+    if (popupClicked) return;
+    popupClicked = true;
+    setTimeout(function() {
+      popupClicked = false;
+    }, POPUP_DOUBLE_CLICK_TIMEOUT);
     /*
       Check if popup already exists
     */
@@ -49,11 +55,10 @@ function add_popup(layerID) {
       Construct Coordinates From Event
     */
     const coordinates = createCoordinates(event);
-    console.log(coordinates)
-    console.log(event.features[0].properties)
 
     const featureID = event.features[0].properties.id;
     const lookup = map_store.cold.currentChoropleth === null ? layerID : map_store.cold.currentChoropleth;
+
     $.ajax({
       type: "GET",
       url: `/popup/${lookup}/${featureID}?lang=en`,
@@ -64,91 +69,15 @@ function add_popup(layerID) {
         const {html} = data;
         popup.innerHTML = html;
 
+        new maplibregl.Popup({
+          // https://maplibre.org/maplibre-gl-js-docs/api/markers/#popup-parameters
+          maxWidth: "280px",
+        }).setLngLat(coordinates).setHTML(popup.innerHTML).addTo(map);
+
         if ("chart" in data) {
-          // Chart Title
-          const {chart: {title}} = data;
-
-          // Chart
-          const chartElement = popup.querySelector("#js-popup__chart");
-          const chart = echarts.init(chartElement, null, {renderer: 'svg'});
-          // TODO: use lookup property in payload to construct chart dynamically. For now we assume bar chart type.
-          // TODO: In this fetch we always expect one payload item. Make failsafe.
-          const {chart: {series}} = data;
-          const xAxisData = createListByName("key", series[0].data);
-          const yAxisData = createListByName("value", series[0].data);
-          const option = {
-            title: {
-              text: title,
-              textStyle: {
-                color: '#002E50',
-                fontSize: 14,
-                fontWeight: 400,
-                lineHeight: 16
-              },
-              left: 'center'
-            },
-            animation: false,
-            tooltip: {
-              trigger: 'axis',
-              axisPointer: {
-                type: 'shadow'
-              }
-            },
-            grid: {
-              left: 16,
-              right: 0,
-              bottom: 32,
-              top: 48,
-              containLabel: true
-            },
-            textStyle: {
-              color: '#002E50'
-            },
-            xAxis: [{
-              type: 'category',
-              data: xAxisData,
-              axisTick: {
-                show: false
-              },
-              axisLine: {
-                show: true,
-                lineStyle: {
-                  color: '#ECF2F6'
-                }
-              },
-            }],
-            yAxis: [{
-              type: 'value',
-              splitLine: {
-                show: true,
-                lineStyle: {
-                  color: '#ECF2F6'
-                }
-              }
-            }],
-            series: [{
-              name: 'Direct',
-              type: 'line',
-              symbol: 'circle',
-              symbolSize: 6,
-              data: yAxisData,
-              lineStyle: {
-                color: '#002E50'
-              },
-              itemStyle: {
-                color: '#002E50'
-              }
-            }]
-          };
-          chart.setOption(option);
+          // createChart function must be defined in project app (not in mapengine)
+          createChart("js-popup__chart", data.chart);
         }
-
-        requestAnimationFrame(() => {
-            new maplibregl.Popup({
-              // https://maplibre.org/maplibre-gl-js-docs/api/markers/#popup-parameters
-              maxWidth: "280px",
-            }).setLngLat(coordinates).setHTML(popup.innerHTML).addTo(map);
-          });
       }
     });
   });